+++ module/config.js
@@ -379,24 +379,24 @@ HERO.movementPowers = {
 HERO.powers = {
 
     // Characteristics (will likely use active effects for these)
-    "STR": { powerType: ["characteristic"], costEnd: true },
-    "DEX": { powerType: ["characteristic"] },
-    "CON": { powerType: ["characteristic"] },
-    "INT": { powerType: ["characteristic"] },
-    "EGO": { powerType: ["characteristic"] },
-    "PRE": { powerType: ["characteristic"] },
-    "OCV": { powerType: ["characteristic"] },
-    "DCV": { powerType: ["characteristic"] },
-    "OMCV": { powerType: ["characteristic"] },
-    "DMCV": { powerType: ["characteristic"] },
-    "SPD": { powerType: ["characteristic"] },
-    "PD": { powerType: ["characteristic", "defense"] },
-    "ED": { powerType: ["characteristic", "defense"] },
-    "REC": { powerType: ["characteristic"] },
-    "END": { powerType: ["characteristic"] },
-    "BODY": { powerType: ["characteristic"] },
-    "STUN": { powerType: ["characteristic"] },
-    "COM": { powerType: ["characteristic"] },
+    "STR": { powerType: ["characteristic"], duration: "persistent", costEnd: true },
+    "DEX": { powerType: ["characteristic"], duration: "persistent" },
+    "CON": { powerType: ["characteristic"], duration: "persistent" },
+    "INT": { powerType: ["characteristic"], duration: "persistent" },
+    "EGO": { powerType: ["characteristic"], duration: "persistent" },
+    "PRE": { powerType: ["characteristic"], duration: "persistent" },
+    "OCV": { powerType: ["characteristic"], duration: "persistent" },
+    "DCV": { powerType: ["characteristic"], duration: "persistent" },
+    "OMCV": { powerType: ["characteristic"], duration: "persistent" },
+    "DMCV": { powerType: ["characteristic"], duration: "persistent" },
+    "SPD": { powerType: ["characteristic"], duration: "persistent" },
+    "PD": { powerType: ["characteristic", "defense"], duration: "persistent" },
+    "ED": { powerType: ["characteristic", "defense"], duration: "persistent" },
+    "REC": { powerType: ["characteristic"], duration: "persistent" },
+    "END": { powerType: ["characteristic"], duration: "persistent" },
+    "BODY": { powerType: ["characteristic"], duration: "persistent" },
+    "STUN": { powerType: ["characteristic"], duration: "persistent" },
+    "COM": { powerType: ["characteristic"], duration: "persistent" },
 
     // Misc
     "CLINGING": { powerType: ["standard"] },
@@ -515,7 +515,7 @@ HERO.powers = {
         costEnd: false,
         costPerLevel: 0.25,
     },
-    
+
 
 
 
@@ -627,7 +627,7 @@ HERO.powers = {
         costEnd: true,
         costPerLevel: 1,
     },
-    "INVISIBILITY": { 
+    "INVISIBILITY": {
         name: "Invisibility",
         powerType: ["sense-affecting"],
         percievability: "Special",
@@ -641,12 +641,14 @@ HERO.powers = {
     "FORCEWALL": {
         powerType: ["defense"],
         name: "Barrier",
+        duration: "Instant",
         costEnd: true,
         costPerLevel: 3 / 2,
     }, // AKA BARRIER
     "FORCEFIELD": {
         powerType: ["defense"],
         name: "Resistant Protection",
+        duration: "Persistent",
         costPerLevel: 3 / 2
     },  // AKA RESISTANT PROTECTION
     "FLASHDEFENSE": {
@@ -662,6 +664,11 @@ HERO.powers = {
     "MENTALDEFENSE": {
         powerType: ["defense", "special"],
         name: "Mental Defense",
+        percievability: "Imperceptible",
+        target: "self only",
+        range: "self",
+        costEnd: false,
+        duration: "Persistent",
         costPerLevel: 1
     },
     "POWERDEFENSE": {
@@ -746,7 +753,10 @@ HERO.powers = {
     },
 
     // Powers can include Talents
-    "COMBAT_LUCK": { powerType: ["talent", "defense"], costPerLevel: 6 },
+    "COMBAT_LUCK": { 
+        powerType: ["talent", "defense"], 
+        duration: "constant",  // behaves like a constant power
+        costPerLevel: 6 },
     "COMBAT_SENSE": {
         powerType: ["talent"],
         name: "Combat Sense",
@@ -842,6 +852,7 @@ HERO.powers = {
         costPerLevel: 1 / 5,
     },
     "CUSTOMPOWER": { powerType: ["custom"] },
+    "naturalBodyHealing": {},
 }
 
 HERO.powers5e = {
@@ -869,6 +880,7 @@ HERO.powers5e = {
     "ARMOR": {
         powerType: ["defense"],
         name: "Resistant Protection",
+        duration: "Persistent",
         costPerLevel: 3 / 2
     },  // AKA RESISTANT PROTECTION
     "DAMAGERESISTANCE": {
@@ -884,6 +896,7 @@ HERO.powers5e = {
     "FORCEFIELD": {
         powerType: ["defense"],
         name: "Resistant Protection",
+        duration: "Constant",
         costEnd: true,
         costPerLevel: 3 / 2
     },
@@ -910,7 +923,7 @@ HERO.ModifierOverride = {
     "DIMENSIONS": { BASECOST: 5 },
     "DOESBODY": { dc: true },
     "DOUBLEKB": { dc: true },
-    "ENDURANCERESERVEREC": {BASECOST: 2/3},
+    "ENDURANCERESERVEREC": { BASECOST: 2 / 3 },
     "ENERGY": { BASECOST: 5 }, // DAMAGENEGATION
     "HARDENED": { BASECOST: 0.25 },
     "IMPENETRABLE": { BASECOST: 0.25 },