+++ module/item/item-attack.js
@@ -66,7 +66,7 @@ export async function AttackOptions(item) {
         if (combatants && typeof dragRuler != 'undefined') {
 
             if (tokens.length === 1) {
-                
+
                 let distance = dragRuler.getMovedDistanceFromToken(token);
                 let speed = dragRuler.getRangesFromSpeedProvider(token)[1].range;
                 let delta = distance;
@@ -74,19 +74,7 @@ export async function AttackOptions(item) {
                     delta = speed - delta;
                 }
                 data.velocity = delta * 5;
-                // velocityDC = Math.floor(velocity / 10);
-                // if (velocityDC > 0) {
-                //     dc += velocityDC;
-                //     let title = `Started at 0 velocity.`;
-                //     if (delta === distance) {
-                //         title += `Moved ${distance}m.`;
-                //     }
-                //     if (delta != distance) {
-                //         title += ` Increasing velocity over ${speed / 2}m, then decreasing velocity.`;
-                //     }
-                //     title += `<br>Expected to move ${delta}m+ before end of phase to stop at 0 velocity (completing the Move By)`;
-                //     tags.push({ value: `${velocityDC}DC`, name: 'Velocity', title: title })
-                // }
+
             }
         }
 
@@ -110,6 +98,23 @@ export async function AttackOptions(item) {
     if (game.settings.get("hero6efoundryvttv2", "hit locations") && !item.system.noHitLocations) {
         data.useHitLoc = true;
         data.hitLoc = CONFIG.HERO.hitLocations;
+    }
+
+    // Combat Skill Levels
+    const csl = CombatSkillLevelsForAttack(item);
+    if (csl && csl.skill) {
+        let _ocv = csl.omcv > 0 ? 'omcv' : 'ocv';
+        data.cslChoices = { [_ocv]: _ocv };
+        if (csl.skill.system.OPTION != "SINGLE") {
+            data.cslChoices.dcv = "dcv";
+            data.cslChoices.dc = "dc";
+        }
+
+        // CSL radioBoxes names
+        data.csl = []
+        for (let c = 0; c < parseInt(csl.skill.system.LEVELS.value); c++) {
+            data.csl.push({ name: `system.csl.${c}`, value: csl.skill.system.csl[c] })
+        }
     }
 
     const template = "systems/hero6efoundryvttv2/templates/attack/item-attack-card.hbs"
@@ -146,6 +151,28 @@ async function _processAttackOptions(item, form) {
         options[key] = value
     }
 
+    // Combat Skill Levels (update SKILL if changed)
+    const csl = CombatSkillLevelsForAttack(item);
+    const checked = form.querySelectorAll(".combat-skill-levels input:checked");
+    if (csl && checked) {
+        let updateRequired = false;
+        for (let input of checked) {
+            let m = input.name.match(/\.(\w+)\.(\d+)/);
+            let name = m[1];
+            let idx = m[2];
+
+            if (csl.skill.system.csl[idx] != input.value) {
+                csl.skill.system.csl[idx] = input.value;
+                updateRequired = true;
+            }
+        }
+        if (updateRequired) {
+            await csl.skill.update({ 'system.csl': csl.skill.system.csl });
+        }
+    }
+
+
+
     await AttackToHit(item, options)
 }
 
@@ -181,10 +208,6 @@ export async function AttackToHit(item, options) {
         rollEquation = modifyRollEquation(rollEquation, ocvMod);
         tags.push({ value: ocvMod, name: item.name })
     }
-
-    // const autoMod = parseInt(item.actor.system.characteristics.ocv.autoMod) || 0
-    // if (autoMod != 0) {
-    //     rollEquation = modifyRollEquation(rollEquation, autoMod);
 
     // Set +1 OCV
     const setManeuver = item.actor.items.find(o => o.type == 'maneuver' && o.name === 'Set' && o.system.active)
@@ -230,24 +253,39 @@ export async function AttackToHit(item, options) {
     }
 
     let dcv = parseInt(item.system.dcv) + csl.dcv
+
+    // Haymaker -5 DCV
+    const haymakerManeuver = item.actor.items.find(o => o.type == 'maneuver' && o.name === 'Haymaker' && o.system.active)
+    if (haymakerManeuver) {
+        dcv -= 4;
+    }
+
     if (dcv != 0) {
 
         // Make sure we don't already have this activeEffect
         let prevActiveEffect = Array.from(item.actor.allApplicableEffects()).find(o => o.origin === item.uuid);
         if (!prevActiveEffect) {
+
+            // Estimate of how many seconds the DCV penalty lasts (until next phase).
+            // In combat.js#_onStartTurn we remove this AE for exact timing.
+            let seconds = Math.ceil(12 / parseInt(item.actor.system.characteristics.spd.value));
+
             let activeEffect = {
                 label: `${item.name} ${("+" + dcv).replace("+-", "-")} DCV`,
-                icon: "icons/svg/downgrade.svg",
-                origin: item.uuid,
+                icon: (dcv) < 0 ? "icons/svg/downgrade.svg" : "icons/svg/upgrade.svg",
                 changes: [
                     { key: "system.characteristics.dcv.value", value: dcv, mode: CONST.ACTIVE_EFFECT_MODES.ADD },
                 ],
+                origin: item.uuid,
                 duration: {
-                    type: "nextPhase"
+                    seconds: seconds,
                 },
-                transfer: true,
+                flags: {
+                    nextPhase: true,
+                }
             }
-            //await item.actor.createEmbeddedDocuments("ActiveEffect", [activeEffect]);
+            //await item.addActiveEffect(activeEffect);
+            await item.actor.createEmbeddedDocuments("ActiveEffect", [activeEffect]);
         }
 
     }
@@ -474,46 +512,6 @@ export async function _onRollDamage(event) {
 
     let damageRoll = convertFromDC(item, dc); //(item.system.dice === 0) ? "" : item.system.dice + "d6";
 
-    //let tags = []
-
-    // BASE ATTACK
-    // let baseTag = ""
-    // if (parseInt(item.system.dice) > 0) {
-    //     //tags.push({ value: item.system.dice + "d6", name: "base" })
-    //     baseTag = item.system.dice + "d6";
-    // }
-    // const extraDiceDamage = determineExtraDiceDamage(item)
-    // if (extraDiceDamage !== "") {
-    //     //tags.push({ value: extraDiceDamage, name: "extraDice" })
-    //     damageRoll += extraDiceDamage
-    //     baseTag += extraDiceDamage;
-    // }
-    // tags.push({ value: baseTag || 0, name: item.name })
-
-    // const strDamage = determineStrengthDamage(item, toHitData.effectivestr)
-    // if (strDamage) {
-    //     tags.push({ value: damageRollToTag(strDamage), name: "strength" })
-    //     damageRoll += strDamage
-    // }
-
-
-
-    // const csl = CombatSkillLevelsForAttack(item)
-    // if (csl && csl.dc > 0) {
-
-    //     let cslDamage = csl.dc + "d6"
-    //     if (item.system.killing) {
-    //         cslDamage = Math.floor(csl.dc / 3) + "d6";
-    //         if (csl.dc % 3 >= 0.5) {
-    //             cslDamage += " + 1d3"
-    //         } else if (csl.dc % 3 >= 0.2) {
-    //             cslDamage += " + 1"
-    //         }
-    //     }
-
-    //     tags.push({ value: cslDamage, name: csl.item.name })
-    //     damageRoll += cslDamage
-    // }
 
     damageRoll = simplifyDamageRoll(damageRoll)
 
@@ -635,6 +633,9 @@ export async function _onApplyDamageToSpecificToken(event, tokenId) {
         return ui.notifications.error(`Attack details are no longer availble.`);
     }
     const actor = item.actor
+
+
+
     const template = "systems/hero6efoundryvttv2/templates/chat/apply-damage-card.hbs"
 
     const token = canvas.tokens.get(tokenId)
@@ -642,6 +643,8 @@ export async function _onApplyDamageToSpecificToken(event, tokenId) {
     if (!token) {
         return ui.notifications.warn(`You must select at least one token before applying damage.`);
     }
+
+
 
     // Spoof previous roll (foundry won't process a generic term, needs to be a proper Die instance)
     let newTerms = JSON.parse(damageData.terms);
@@ -661,15 +664,113 @@ export async function _onApplyDamageToSpecificToken(event, tokenId) {
     }
     let newRoll = Roll.fromTerms(newTerms)
 
-
-
     let automation = game.settings.get("hero6efoundryvttv2", "automation");
+
+
+
+
+
+    // Check for conditional defenses
+    let ignoreDefenseIds = []
+    const conditionalDefenses = token.actor.items.filter(o => (o.system.subType || o.system.type) === "defense" &&
+        (o.system.active || o.effects.find(o => true)?.disabled === false) &&
+        o.system.modifiers.find(p => ["ONLYAGAINSTLIMITEDTYPE", "CONDITIONALPOWER"].includes(p.XMLID))
+    )
+    if (conditionalDefenses.length > 0 && !["AID"].includes(item.system.XMLID)) {
+        const template2 = "systems/hero6efoundryvttv2/templates/attack/item-conditional-defense-card.hbs"
+
+        let options = [];
+        for (let defense of conditionalDefenses) {
+            let option = { id: defense.id, name: defense.name, checked: true, conditions: "" }
+            // for (let modifier of defense.system.modifiers.filter(p => ["ONLYAGAINSTLIMITEDTYPE", "CONDITIONALPOWER"].includes(p.XMLID))) {
+            //     option.conditions += modifier.OPTION_ALIAS;
+            //     if (modifier.COMMENTS) {
+            //         option.conditions += ` (${modifier.COMMENTS})`;
+            //     }
+            //     option.conditions += ". ";
+            // }
+            option.description = defense.system.description;
+            options.push(option);
+        }
+
+        let data = {
+            token,
+            item,
+            conditionalDefenses: options,
+        }
+
+        const html = await renderTemplate(template2, data)
+
+        //let cancelled = true;
+
+        async function getDialogOutput() {
+            return new Promise(resolve => {
+                const dataConditionalDefenses = {
+                    title: item.actor.name + " conditional defenses",
+                    content: html,
+                    buttons: {
+                        normal: {
+                            label: "Apply Damage",
+                            callback: (html) => { resolve(html.find("form input")) }
+                            // async function (html) {
+                            // //cancelled = false;
+                            // let inputs = html.find("form input");
+                            // return inputs;
+                            // // for (let input of inputs) {
+                            // //     if (input.checked) {
+                            // //         ignoreDefenseIds.push(input.id);
+                            // //     }
+                            // // }
+                        },
+                        cancel: {
+                            label: "cancel",
+                            callback: () => { resolve(null) }
+                        }
+                    },
+                    default: "normal",
+                    close: () => { resolve(null) }
+                }
+                new Dialog(dataConditionalDefenses).render(true)
+            });
+        }
+
+        const inputs = await getDialogOutput();
+        if (inputs === null) return;
+
+        let names = [];
+        for (let input of inputs) {
+            if (!input.checked) {
+                ignoreDefenseIds.push(input.id);
+                names.push(token.actor.items.get(input.id).name);
+            }
+        }
+
+        if (names.length > 0) {
+            let content = `The following defenses were not applied to ${token.actor.name}:<ul>`;
+            for (let name of names) {
+                content += `<li>${name}</li>`
+            }
+            content += "</ul>";
+
+            const speaker = ChatMessage.getSpeaker({ actor: token.actor })
+            speaker["alias"] = token.actor.name
+            const chatData = {
+                user: game.user._id,
+                type: CONST.CHAT_MESSAGE_TYPES.OTHER,
+                content,
+                whisper: ChatMessage.getWhisperRecipients("GM"),
+                speaker,
+            }
+
+            await ChatMessage.create(chatData)
+        }
+    }
 
     // -------------------------------------------------
     // determine active defenses
     // -------------------------------------------------
     let defense = "";
-    let [defenseValue, resistantValue, impenetrableValue, damageReductionValue, damageNegationValue, knockbackResistance, defenseTags] = determineDefense(token.actor, item)
+    let [defenseValue, resistantValue, impenetrableValue, damageReductionValue, damageNegationValue, knockbackResistance, defenseTags] = determineDefense(token.actor, item, { ignoreDefenseIds })
     if (damageNegationValue > 0) {
         defense += "Damage Negation " + damageNegationValue + "DC(s); "
     }